update list tools, move akshara class to __init__
